backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic[email]==2.9.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
cricheroes==1.0.10
razorpay==1.4.2
websockets==13.1
aiofiles==24.1.0
pandas==2.1.4
openpyxl==3.1.2
